Create version1
We first applied for the access for using twitter api and created our app. Then we learned about how to use the api from online resources. This is the first code we put on and it is pretty simple. We could use it to access the stream and we can use the filter to select particular user that we want to listen. The information we get is in json. We will learn more and improve the code very soon.