Create Version1
We first applied for the access of using twitter api and created our own app. Then we learned how to use api and finished our first code for the project. We can use it to listen to a particular user, we use Justin Bieber as an example. The information we got is in json.